:sparkles:新增单独对游戏中的非配置语言的导入

+++ Excel2Bytes/GenerateUtil.py
@@ -5,7 +5,7 @@ from JsonUtil import InitTableJsonData, LoadTableJsonData
 from ExcelUtil import CopyScripts, CopyBytes, DeleteScripts, DeleteBytes
 from FieldGenerate import GenerateFieldBytes, GenerateLNGBytes, GenerateFindFieldBytes, GenerateNoExportLNGBytes
 from LanguageUtil import InitLanguage, SaveLanguage
-from GlobalUtil import InitFileDir, TablePath, GenerateScriptType
+from GlobalUtil import InitFileDir, TablePath, GenerateScriptType, IsUpdateAllLNG
 from ResRefUtil import SaveResList
 
 
@@ -38,9 +38,10 @@ def GenerateBytes(scriptType, excelPath, sheetName, scriptName, extraNamespace):
         GenerateNoExportLNGBytes(excelPath, sheetName)
 
 
-def ExportAllData():
-    DeleteFile()
-    tableConfig = LoadTableJsonData()
+def ExportData(tableConfig, isUpdateAllLNG=False, isDeleteFile=False):
+    IsUpdateAllLNG = isUpdateAllLNG
+    if IsUpdateAllLNG:
+        DeleteFile()
     for excelItem in tableConfig.items():
         for sheetItem in excelItem[1].items():
             GenerateBytes(sheetItem[1]['ImportType'], os.path.join(TablePath, excelItem[0]), sheetItem[0],
@@ -49,3 +50,23 @@ def ExportAllData():
     SaveResList()
     # CopyExportFiles()
     # CreateTableManagerCs()
+
+
+def ExportAllData():
+    tableConfig = LoadTableJsonData()
+    ExportData(tableConfig, True, True)
+
+
+def ExportExtraLNG():
+    tableConfig = LoadTableJsonData()
+    config = {}
+    for excelItem in tableConfig.items():
+        for sheetItem in excelItem[1].items():
+            importType = sheetItem[1]['ImportType']
+            if importType == GenerateScriptType.NoExportLNGType.name or importType == GenerateScriptType.LNGType.name:
+                if excelItem[0] not in config:
+                    config[excelItem[0]] = {sheetItem[0]: sheetItem[1]}
+                else:
+                    config[excelItem[0]][sheetItem[0]] = sheetItem[1]
+    print(config)
+    ExportData(config, True, False)
